created controllers and routes with validation
