+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^0.21.1",
     "bcrypt": "^5.0.1",
     "express": "^4.17.1",
+    "express-validation": "^3.0.8",
     "pg": "^8.4.0",
     "reflect-metadata": "^0.1.10",
     "typeorm": "0.2.29",
