+++ src/validators/project/ProjectValidators.ts
@@ -1,0 +1,10 @@
+import {Joi} from 'express-validation'
+
+const createProjectrValidator = {
+  body: Joi.object({
+    description: Joi.string().required(),
+    userId: Joi.string().required(),
+  })
+};
+
+export { createProjectrValidator };
